Split install from config to build an image for SeaPac.

+++ config/core_install.sh
@@ -4,7 +4,9 @@
 #
 # Uncomment this statement for debug echos
 DEBUG=1
-myname="`basename $0`"
+
+scriptname="`basename $0`"
+UDR_INSTALL_LOGFILE="/var/log/udr_install.log"
 
 # do upgrade, update outside of script since it can take some time
 UPDATE_NOW=false
@@ -16,7 +18,9 @@ NONESSENTIAL_PKG_LIST="mg jed whois"
 NONESSENTIAL_PKG=true
 
 BUILDTOOLS_PKG_LIST="rsync build-essential autoconf dh-autoreconf automake libtool git libasound2-dev libncurses5-dev"
-# If the following is set to true bluetooth will be disabled
+REMOVE_PKG_LIST="libax25 libax25-dev ax25-apps ax25-tools"
+
+# If the following is set to true, bluetooth will be disabled
 SERIAL_CONSOLE=false
 
 # trap ctrl-c and call function ctrl_c()
@@ -28,7 +32,7 @@ function dbgecho { if [ ! -z "$DEBUG" ] ; then echo "$*"; fi }
 
 function is_pkg_installed() {
 
-return $(dpkg-query -W -f='${Status}' $1 2>/dev/null | grep -c "ok installed")
+return $(dpkg-query -W -f='${Status}' $1 2>/dev/null | grep -c "ok installed" >/dev/null 2>&1)
 }
 
 # ===== function ctrl_c trap handler
@@ -61,7 +65,7 @@ function install_direwolf_source() {
 
 # ===== main
 
-echo "Initial core install/config script"
+echo "Initial core install script"
 
 # Be sure we're running as root
 if [[ $EUID != 0 ]] ; then
@@ -87,8 +91,8 @@ if [ "$NONESSENTIAL_PKG" = "true" ] ; then
    for pkg_name in `echo ${NONESSENTIAL_PKG_LIST}` ; do
 
       is_pkg_installed $pkg_name
-      if [ $? -eq 0 ] ; then
-         echo "$myname: Will Install $pkg_name program"
+      if [ $? -ne 0 ] ; then
+         echo "$scriptname: Will Install $pkg_name program"
          needs_pkg=true
          break
       fi
@@ -115,8 +119,8 @@ needs_pkg=false
 for pkg_name in `echo ${BUILDTOOLS_PKG_LIST}` ; do
 
    is_pkg_installed $pkg_name
-   if [ $? -eq 0 ] ; then
-      echo "$myname: Will Install $pkg_name program"
+   if [ $? -ne 0 ] ; then
+      echo "$scriptname: Will Install $pkg_name program"
       needs_pkg=true
       break
    fi
@@ -141,90 +145,6 @@ if [ ! -d /lib/modules/$(uname -r)/ ] ; then
    echo "shutdown -r now"
    echo "and log back in"
    exit 1
-fi
-
-echo " === Verify not using default password"
-# is there even a user pi?
-ls /home | grep pi > /dev/null 2>&1
-if [ $? -eq 0 ] ; then
-   echo "User pi found"
-   echo "Determine if default password is being used"
-
-   # get salt
-   SALT=$(grep -i pi /etc/shadow | awk -F\$ '{print $3}')
-
-   PASSGEN=$(mkpasswd --method=sha-512 --salt=$SALT raspberry)
-   PASSFILE=$(grep -i pi /etc/shadow | cut -d ':' -f2)
-
-#   dbgecho "SALT: $SALT"
-#   dbgecho "pass file: $PASSFILE"
-#   dbgecho "pass  gen: $PASSGEN"
-
-   if [ "$PASSFILE" = "$PASSGEN" ] ; then
-      echo "User pi is using default password"
-      echo "Need to change your password for user pi NOW"
-      read -t 1 -n 10000 discard
-      passwd pi
-      if [ $? -ne 0 ] ; then
-         echo "Failed to set password, exiting"
-	 exit 1
-      fi
-   else
-      echo "User pi not using default password."
-   fi
-
-else
-   echo "User pi NOT found"
-fi
-
-# Check hostname
-echo " === Verify hostname"
-HOSTNAME=$(cat /etc/hostname | tail -1)
-dbgecho "Current hostname: $HOSTNAME"
-
-if [ "$HOSTNAME" = "raspberrypi" ] || [ "$HOSTNAME" = "compass" ] ; then
-   # Change hostname
-   echo "Using default host name: $HOSTNAME, change it"
-   echo "Enter new host name followed by [enter]:"
-   read -t 1 -n 10000 discard
-   read -e HOSTNAME
-   echo "$HOSTNAME" > /etc/hostname
-fi
-
-# Get hostname again incase it was changed
-HOSTNAME=$(cat /etc/hostname | tail -1)
-
-# Be sure system host name can be resolved
-
-grep "127.0.1.1" /etc/hosts
-if [ $? -eq 0 ] ; then
-   # Found 127.0.1.1 entry
-   # Be sure hostnames match
-   HOSTNAME_CHECK=$(grep "127.0.1.1" /etc/hosts | awk {'print $2'})
-   if [ "$HOSTNAME" != "$HOSTNAME_CHECK" ] ; then
-      echo "Make host names match between /etc/hostname & /etc/hosts"
-      sed -i -e "/127.0.1.1/ s/127.0.1.1\t.*/127.0.1.1\t$HOSTNAME/" /etc/hosts
-   else
-      echo "host names match between /etc/hostname & /etc/hosts"
-   fi
-else
-   # Add a 127.0.1.1 entry to /etc/hosts
-   sed -i '1i\'"127.0.1.1\t$HOSTNAME $HOSTNAME.localnet" /etc/hosts
-   if [ $? -ne 0 ] ; then
-      echo "Failed to modify /etchosts file"
-   fi
-fi
-
-DATETZ=$(date +%Z)
-dbgecho "Time zone: $DATETZ"
-
-if [ "$DATETZ" == "UTC" ] ; then
-   echo " === Set time zone"
-   echo " ie. select America, then scroll down to 'Los Angeles'"
-   echo " then hit tab & return ... wait for it"
-   # pause to read above msg
-   sleep 4
-   dpkg-reconfigure tzdata
 fi
 
 echo " === Modify /boot/config.txt"
@@ -275,6 +195,24 @@ if [ $? -ne 0 ] ; then
 fi
 
 echo " === Install libax25, ax25apps & ax25tools"
+# libax25, ax25apps & ax25tools are about to be installed from source
+# - first check that any packages are installed and uninstall them
+echo " Check for previous packages installed"
+
+for pkg_name in `echo ${REMOVE_PKG_LIST}` ; do
+
+   is_pkg_installed $pkg_name
+   if [ $? -eq 0 ] ; then
+      echo "$scriptname: Will remove $pkg_name"
+      apt-get purge -y -q $pkg_name
+      if [ "$?" -ne 0 ] ; then
+         echo "Conflicting package removal failed. Please try this command manually:"
+         echo "apt-get purge -y $pkg_name"
+      fi
+   fi
+done
+
+echo "Begin building libax25, ax25apps & ax25tools "
 
 # Does source directory for ax25 utils exist?
 SRC_DIR="/usr/local/src/ax25"
@@ -367,8 +305,13 @@ fi
 
 # Since libax25 was built from source
 # need to add a symbolic link to the /usr/lib directory
-
-ln -s /usr/local/lib/libax25.so.1 /usr/lib/libax25.so.0
+filename=/usr/lib/libax25.so.0
+if [[ -L "$filename" ]] ; then
+   echo " Symbolic link to $filename ALREADY exists"
+else
+   echo "Making symbolic link to $filename"
+   ln -s /usr/local/lib/libax25.so.1 /usr/lib/libax25.so.0
+fi
 
 # pkg installed libraries are installed in /usr/lib
 # built libraries are installed in /usr/local/lib
@@ -400,39 +343,15 @@ fi
 
 if [ ! -e /etc/direwolf.conf ] ; then
    echo "Direwolf: NO config file found!!"
-   echo "$myname: direwolf install failed!"
+   echo "$scriptname: direwolf install failed!"
    exit 1
 else
    echo "direwolf: config file found OK"
 fi
 
-echo "=== Set alsa levels for UDRC"
-# Does source directory for udrc alsa level setup script exist?
-SRC_DIR="/usr/local/src/udrc"
-
-if [ ! -d $SRC_DIR ] ; then
-   mkdir -p $SRC_DIR
-   if [ $? -ne 0 ] ; then
-      echo "Problems creating source directory: $SRC_DIR"
-      exit 1
-   fi
-else
-   dbgecho "Source dir: $SRC_DIR already exists"
-fi
-cd $SRC_DIR
-wget -O set-udrc-din6.sh -qt 3 https://goo.gl/7rXUFJ
-if [ $? -ne 0 ] ; then
-   echo "FAILED to download alsa level setup file."
-   exit 1
-fi
-chmod +x set-udrc-din6.sh
-./set-udrc-din6.sh  > /dev/null 2>&1
-
-UDR_INSTALL_LOGFILE="/var/log/udr_install.log"
 echo "$(date "+%Y %m %d %T %Z"): core install script FINISHED" >> $UDR_INSTALL_LOGFILE
 echo
 echo "core install script FINISHED"
 echo
 cd $START_DIR
 /bin/bash $START_DIR/app_install.sh core
-exit 0
